+++ Artifact/Rendering/ParticleSystem.h
@@ -15,7 +15,7 @@ namespace Artifact
 	public:
 		/// <summary>Initializes a new instance of the <see cref="ParticleSystem"/> class.</summary>
 		/// <param name="a_EntitySystem">The entity system to operate on, so that it can fetch the needed components </param>
-		/// <param name="a_MessagingSystem">The a_ messaging system.</param>
+		/// <param name="a_MessagingSystem">The messaging system to communicate with other systems.</param>
 		ParticleSystem(EntitySystem& a_EntitySystem, MessagingSystem& a_MessagingSystem);
 
 		/// <summary>Registers the listeners for the messages it will listen to </summary>
@@ -26,7 +26,7 @@ namespace Artifact
 		
 		/// <summary>Updates the particle simulations</summary>
 		/// <param name="a_DeltaTime">The tiem since hte last update</param>
-		void updateParticleSimulations(float a_DeltaTime) const;
+		void updateParticleSimulations(double a_DeltaTime) const;
 
 		/// <summary>Integrates the attributes of each particle using their corresponding emitter </summary>
 		void integrateAttributes(ComponentHandle<ParticleEmitter> a_ParticleEmitter) const;
